Move `protocol_initializer` database tests into a `tests` module in a separate folder
