+++ mithril-signer/src/database/repository/protocol_initializer_repository.rs
@@ -92,217 +92,3 @@ impl ProtocolInitializerStorer for ProtocolInitializerRepository {
             .collect())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::database::test_helper::{main_db_connection, FakeStoreAdapter};
-    use mithril_common::test_utils::fake_data;
-    use mithril_persistence::sqlite::{ConnectionBuilder, ConnectionOptions};
-
-    use super::*;
-
-    fn setup_protocol_initializers(nb_epoch: u64) -> Vec<(Epoch, ProtocolInitializer)> {
-        let mut values: Vec<(Epoch, ProtocolInitializer)> = Vec::new();
-        for epoch in 1..=nb_epoch {
-            let stake = (epoch + 1) * 100;
-            let protocol_initializer = fake_data::protocol_initializer("1", stake);
-            values.push((Epoch(epoch), protocol_initializer));
-        }
-        values
-    }
-
-    async fn init_store(
-        nb_epoch: u64,
-        retention_limit: Option<u64>,
-    ) -> ProtocolInitializerRepository {
-        let store = ProtocolInitializerRepository::new(
-            Arc::new(main_db_connection().unwrap()),
-            retention_limit,
-        );
-
-        let values = setup_protocol_initializers(nb_epoch);
-        store_protocol_initializers(&store, &values).await;
-
-        store
-    }
-
-    async fn store_protocol_initializers(
-        store: &ProtocolInitializerRepository,
-        values: &[(Epoch, ProtocolInitializer)],
-    ) {
-        for value in values.iter() {
-            store
-                .save_protocol_initializer(value.0, value.1.clone())
-                .await
-                .unwrap();
-        }
-    }
-
-    #[tokio::test]
-    async fn save_key_in_empty_store_return_none_as_previous_value() {
-        let protocol_initializers = setup_protocol_initializers(1);
-
-        let store = init_store(0, None).await;
-        let res = store
-            .save_protocol_initializer(
-                protocol_initializers[0].0,
-                protocol_initializers[0].1.clone(),
-            )
-            .await
-            .unwrap();
-
-        assert!(res.is_none());
-    }
-
-    #[tokio::test]
-    async fn update_protocol_initializer_in_store_return_previous_value() {
-        let protocol_initializers = setup_protocol_initializers(2);
-        let store = init_store(0, None).await;
-        store_protocol_initializers(&store, &protocol_initializers[0..1]).await;
-
-        let res = store
-            .save_protocol_initializer(
-                protocol_initializers[0].0,
-                protocol_initializers[1].1.clone(),
-            )
-            .await
-            .unwrap();
-
-        assert!(res.is_some());
-        assert_eq!(
-            protocol_initializers[0].1.get_stake(),
-            res.unwrap().get_stake()
-        );
-    }
-
-    #[tokio::test]
-    async fn get_protocol_initializer_for_empty_epoch() {
-        let store = init_store(2, None).await;
-        let res = store.get_protocol_initializer(Epoch(0)).await.unwrap();
-
-        assert!(res.is_none());
-    }
-
-    #[tokio::test]
-    async fn get_protocol_initializer_for_existing_epoch() {
-        let store = init_store(2, None).await;
-        let res = store.get_protocol_initializer(Epoch(1)).await.unwrap();
-
-        assert!(res.is_some());
-    }
-
-    #[tokio::test]
-    async fn get_last_protocol_initializer_return_last_one_first() {
-        let store = init_store(0, None).await;
-        let values = setup_protocol_initializers(10);
-        store_protocol_initializers(&store, &values).await;
-
-        let res = store.get_last_protocol_initializer(3).await.unwrap();
-
-        assert_eq!(3, res.len());
-        assert_eq!(values[9].0, res[0].0);
-        assert_eq!(values[8].0, res[1].0);
-        assert_eq!(values[7].0, res[2].0);
-    }
-
-    #[tokio::test]
-    async fn get_last_protocol_initializer_return_all_when_too_few_records() {
-        let store = init_store(0, None).await;
-        let values = setup_protocol_initializers(2);
-        store_protocol_initializers(&store, &values).await;
-
-        let res = store.get_last_protocol_initializer(3).await.unwrap();
-
-        assert_eq!(2, res.len());
-        assert_eq!(values[1].0, res[0].0);
-        assert_eq!(values[0].0, res[1].0);
-    }
-
-    #[tokio::test]
-    async fn prune_epoch_older_than_threshold() {
-        const PROTOCOL_INITIALIZER_PRUNE_EPOCH_THRESHOLD: u64 = 10;
-
-        let store = init_store(0, Some(PROTOCOL_INITIALIZER_PRUNE_EPOCH_THRESHOLD)).await;
-        let values = setup_protocol_initializers(2);
-        store_protocol_initializers(&store, &values).await;
-
-        store
-            .prune(Epoch(2) + PROTOCOL_INITIALIZER_PRUNE_EPOCH_THRESHOLD)
-            .await
-            .unwrap();
-
-        let res = store.get_last_protocol_initializer(10).await.unwrap();
-
-        assert_eq!(1, res.len());
-        assert_eq!(Epoch(2), res[0].0);
-    }
-
-    #[tokio::test]
-    async fn without_threshold_nothing_is_pruned() {
-        let store = init_store(0, None).await;
-        let values = setup_protocol_initializers(2);
-        store_protocol_initializers(&store, &values).await;
-
-        store.prune(Epoch(100)).await.unwrap();
-
-        let res = store.get_last_protocol_initializer(10).await.unwrap();
-        assert_eq!(2, res.len());
-    }
-
-    #[tokio::test]
-    async fn should_migrate_data_from_adapter() {
-        let migrations = crate::database::migration::get_migrations();
-
-        // TODO: Do it in test_helper (it is done by build_main_db_connection)
-        fn create_connection_builder() -> ConnectionBuilder {
-            ConnectionBuilder::open_memory()
-                .with_options(&[ConnectionOptions::ForceDisableForeignKeys])
-        }
-        let connection = Arc::new(create_connection_builder().build().unwrap());
-        let protocol_initializer_adapter =
-            FakeStoreAdapter::new(connection.clone(), "protocol_initializer");
-        // The adapter will create the table.
-        protocol_initializer_adapter.create_table();
-
-        assert!(connection
-            .prepare("select key_hash from protocol_initializer;")
-            .is_ok());
-
-        // Here we can add some data with the old schema.
-        let (_, protocol_initializer_to_retrieve) = &setup_protocol_initializers(1)[0];
-
-        assert!(!protocol_initializer_adapter.is_key_hash_exist("HashEpoch5"));
-
-        protocol_initializer_adapter
-            .store_record("HashEpoch5", &Epoch(5), protocol_initializer_to_retrieve)
-            .unwrap();
-
-        assert!(protocol_initializer_adapter.is_key_hash_exist("HashEpoch5"));
-
-        // We finish the migration
-        create_connection_builder()
-            .apply_migrations(&connection, migrations)
-            .unwrap();
-
-        assert!(connection
-            .prepare("select key_hash from protocol_initializer;")
-            .is_err());
-        assert!(connection
-            .prepare("select * from protocol_initializer;")
-            .is_ok());
-
-        let value: i64 = connection
-            .query_single_cell("select count(*) from protocol_initializer", &[])
-            .unwrap();
-        assert_eq!(value, 1);
-
-        // We can check that data are migrated.
-        let store = ProtocolInitializerRepository::new(connection, None);
-        let protocol_initializer = store.get_protocol_initializer(Epoch(5)).await.unwrap();
-
-        assert_eq!(
-            protocol_initializer.unwrap().get_stake(),
-            protocol_initializer_to_retrieve.get_stake()
-        );
-    }
-}
