+++ mithril-signer/src/database/tests/mod.rs
@@ -1,0 +1,1 @@
+mod protocol_initializer;
